backend/requirements.txt
Flask==2.3.2
Flask-CORS==3.0.10
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
